+++ legacy/edje/src/lib/edje_text.c
@@ -266,7 +266,7 @@ static const char *
 _edje_text_font_get(const char *base, const char *new, char **free_later)
 {
    const char *base_style, *new_style, *aux;
-   int font_len, style_len;
+   size_t font_len, style_len;
 
    if (base && (!new))
      return base;
@@ -283,7 +283,7 @@ _edje_text_font_get(const char *base, const char *new, char **free_later)
 
    font_len = strlen(new);
    aux = strchr(base_style, ',');
-   style_len = (aux) ? (aux - base_style) : (int) strlen(base_style);
+   style_len = (aux) ? (aux - base_style) : strlen(base_style);
 
    *free_later = malloc(font_len + style_len + 1);
    memcpy(*free_later, new, font_len);
@@ -365,7 +365,7 @@ _edje_text_recalc_apply(Edje *ed, Edje_Real_Part *ep,
 	  {
              char *font2;
              
-             int len = strlen(font) + sizeof("edje/fonts/") + 1;
+             size_t len = strlen(font) + sizeof("edje/fonts/") + 1;
              font2 = alloca(len);
              sprintf(font2, "edje/fonts/%s", font);
              font = font2;
